Launcher3: Fix launcher preview rendering
Launcher preview rendering fails on AOSP Launcher3 because the search
bar (AppsSearchContainerLayout) tries to find BaseDraggingActivity
during inflation, which doesn't exist in the preview hierarchy.

Remove the search bar from the launcher preview layout entirely to fix
the issue. This is actually more consistent with our real launcher
appearance because it doesn't have a search bar to begin with.

Change-Id: Ic549fd9d592bc702e0c12285fb016b908e49d514
Signed-off-by: DennySPb <dennyspb@gmail.com>